created ShoppingListActivity with navBar.
