+++ DietScoopApp/app/src/main/java/com/example/dietscoop/Activities/NavigationActivity.java
@@ -1,55 +1,26 @@
 package com.example.dietscoop.Activities;
 
 import android.content.Intent;
-import android.os.Bundle;
 import android.util.Log;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.ImageButton;
 
 import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 import androidx.appcompat.app.ActionBar;
 import androidx.appcompat.app.AppCompatActivity;
 
-import com.example.dietscoop.Data.Recipe.Recipe;
 import com.example.dietscoop.R;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
-import com.google.android.material.navigation.NavigationBarView;
 import com.google.firebase.auth.FirebaseAuth;
 
 public abstract class NavigationActivity extends AppCompatActivity implements BottomNavigationView.OnItemSelectedListener {
 
     BottomNavigationView navBar;
-    ActionBar topBar;
 
     protected void initNavigationActivity() {
         navBar = findViewById(R.id.bottomNavBar);
         navBar.setOnItemSelectedListener(this);
-
-        topBar = getSupportActionBar();
-        topBar.setDisplayOptions(ActionBar.DISPLAY_SHOW_CUSTOM);
-        topBar.setDisplayShowCustomEnabled(true);
-        topBar.setCustomView(R.layout.top_bar_layout);
-
-        View topBarView = topBar.getCustomView();
-
-        ImageButton logout = topBarView.findViewById(R.id.logout_button);
-        ImageButton addItem = topBarView.findViewById(R.id.add_button);
-
-        logout.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                logout();
-            }
-        });
-
-        addItem.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                onAddClicked();
-            }
-        });
     }
 
     /**
@@ -107,14 +78,4 @@ public abstract class NavigationActivity extends AppCompatActivity implements Bo
         }
         return true;
     }
-
-    protected void logout() {
-        FirebaseAuth.getInstance().signOut();
-
-        Intent intent = new Intent(this, MainActivity.class);
-        startActivity(intent);
-
-    }
-
-    abstract void onAddClicked();
 }
